requirements.txt
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.9
chromadb==0.5.3
pypdf==4.2.0
python-dotenv==1.0.1
tiktoken==0.7.0